Signals can now return objects as responses
